Style list-item::before elements
NOTE: getting and placing  the numbers with JS seems that the code is
more complex, as it should be. This way we can control thenumbers just
by changing the file contents.

The same result can be achieved using the CSS counter function, but
anything that later can make the numbering more complex has to be done
in the code vs adding files to content folder.