Refactor service card button
Added new div that adds the background fill around the button.
Previously the button was handling the background with top and left
borders, and like that could not achieve clear focus state. (Focus could
not get a round shape because of the uneven borders being set).
Now the card and the newly added div element are managing the wave and
the background fill, taking this responsibility off of from the button,
that can have a pleasing and accessible focus state.
The cards still get the distinct background color on button focus, that
is just an added visual enhancement.